# docs/refer/AI4MS-DevPack_v0.3/06_developer_starter/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI4MS API
  version: 0.3.0
  description: Minimal contract for topic scouting, stage agents, editable revisions, human approvals, Stata/Python runs, and evidence workflows.
servers:
  - url: /v1
security:
  - bearerAuth: []
paths:
...
  /runs/{runId}/cancel:
    post:
      operationId: cancelRun
      parameters:
        - name: runId
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        '202': {description: Cancellation requested}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...